refactor: extract/ owns the ABICHECK_* boolean knobs, not model/
Codex review (P1): model/ reads no environment. Its own scoped contract
says it owns shapes and answers "what is this fact", never "how was it
produced" -- so a module that reads the live process environment and
resolves a runtime input does not belong there, and putting it there makes
the innermost ring a general-purpose utility layer that invites
configuration resolution at arbitrary call sites. The reviewer is right,
and the original reason for model/ ("every layer may import it") was an
argument about reachability, not about ownership.

extract/ is a real owner rather than a fallback: every knob in the registry
governs how evidence is captured -- AST-frontend fallback, the castxml
version gate, the system-include probe, the streaming pruner, COMDAT
collection, the preprocessor pre-scan, two-sided extraction concurrency,
and whether the abicheck-cc wrapper captures at all. Every reader is itself
classified extract, except the compare pipeline's concurrency check, and
workflows -> extract is a permitted direction, so this creates no forbidden
edge.

A flat-root leaf with an `accept` disposition was written and then
discarded: the gate only admits a root module as legacy/public/facade, and
recording a brand-new module as "legacy" would have been false. A real
owner existed; the disposition would have been a way of not looking for it.

The module's own docstring records why model/ was wrong, so the next reader
does not repeat the move.
