+++ abicheck/dumper_clang_errors.py
@@ -39,7 +39,7 @@ from . import deadline
 from .dumper_cache import _atomic_copy, _atomic_write_json, ast_memoize_active
 from .dumper_clang_streaming import load_pruned_clang_ast, streaming_prune_suppressed
 from .errors import SnapshotError
-from .model.env_flags import env_flag
+from .extract.env_flags import env_flag
 from .sycl_context import decode_and_select_frontend_context_from_path
 
 log = logging.getLogger(__name__)
